envconfig: accept $DISABLE_CCACHE and $DISABLE_SCCACHE
If $DISABLE_CCACHE is set, do not bother detecting ccache. This aligns
with semantics of this variable in ccache itself.

Similarly, if $DISABLE_SCCACHE is set, do not detect sccache. This
variable has no special meaning for sccache, but using sccache in
general is not always desirable, so there should be a way to prevent its
detection.

Fixes: #2991 #11118