detect_rust_compiler: handle "ccache" correctly in the override case
Similar to !override case, only strip the first word of the command line
if (s)ccache is the command name, not just if it happens to end with
'ccache'.

Thus, 'ccache' will only be stripped if it is a wrapper command
(e. g.: ['ccache', 'gcc', ...]), but not if it is part of the path
(e. g.: ['/usr/lib/ccache/bin/gcc', ...]).